+++ invite_me/model/__init__.py
@@ -1,7 +1,8 @@
 __all__ = ["Request", "Response", "User", "Inviter", "InviterUser"]
 
-from invite_me.model.inviter import Inviter
-from invite_me.model.inviter_user import InviterUser
-from invite_me.model.request import Request
-from invite_me.model.response import Response
-from invite_me.model.user import User
+from .inviter import Inviter
+from .inviter_user import InviterUser
+from .request import Request
+from .response import Response
+from .user import User
+
