+++ invite_me/model/repositories/responses.py
@@ -1,11 +1,27 @@
+from abc import ABC, abstractmethod
 from typing import Dict, Optional
 from uuid import UUID
 
+from sqlalchemy.orm import Session
+
 from invite_me.model import Response
-from invite_me.repositories.interfaces.responses import IResponsesRepository
 
 
-class MockResponsesRepository(IResponsesRepository):
+class ResponsesRepository(ABC):
+    @abstractmethod
+    def create_response(self, response: Response):
+        raise NotImplementedError
+
+
+class SqlAlchemyResponsesRepository(ResponsesRepository):
+    def __init__(self, session: Session):
+        self._session = session
+
+    def create_response(self, response: Response):
+        self._session.add(response)
+
+
+class MockResponsesRepository(ResponsesRepository):
     _store: Dict[UUID, Response]
 
     def __init__(self, store: Optional[Dict[UUID, Response]] = None):
