+++ pyproject.toml
@@ -25,6 +25,7 @@ dependencies = [
     "slack-sdk>=3.33.3",
     "slack-bolt>=1.21.2",
     "alembic>=1.14.0",
+    "mypy>=1.13.0",
 ]
 
 [build-system]
