+++ test.dockerfile
@@ -7,8 +7,7 @@ WORKDIR /app
 RUN uv pip install --system -r pyproject.toml
 
 COPY . /app
-
 WORKDIR /app/invite_me
 
 #CMD ["uv", "run", "celery","-A", "tasks", "worker", "--autoscale", "10", "--loglevel=info"]
-CMD ["uv", "run", "pytest", "-s" ,"tests"]
+CMD ["pytest", "-s" ,"tests"]
