+++ worker.dockerfile
@@ -8,6 +8,9 @@ RUN uv pip install --system -r pyproject.toml
 
 COPY . /app
 
+# need to install 'invite_me' here so celery can call to invite_me functions
+RUN uv pip install --system .
+
 WORKDIR /app/invite_me
 
-CMD ["uv", "run", "celery","-A", "tasks", "worker", "--autoscale", "10", "--loglevel=info"]
+CMD ["celery","-A", "tasks", "worker", "--autoscale", "10", "--loglevel=info"]
